Create Inheritance.java
The Ticket class is the superclass in which the ticket is assigned a number. Also, you get the price of the ticket and print the number of the ticket and the ticket price. 
The WalkupTicket is a sub class of the Ticket class. In this class, we have a constructor “public WalkupTicket(int ticketNumber)” which  takes an integer parameter number, representing the unique number assigned to the ticket. It calls the superclass constructor super(ticketNumber) to initialize the ticket number using the constructor of the superclass Ticket. We also call the methpd getPrice() which returns the ticket price.
AdvanceTicket class extends Ticket and represents tickets purchased in advance of the event. These tickets have varied pricing based on the number of days before the event they were purchased. If purchased 10 or more days in advance, the price is set at $30, otherwise, it's $40. This class implements the getPrice() method to calculate the price accordingly. StudentAdvanceTicket class, which extends AdvanceTicket, further specializes in tickets purchased in advance by students. It inherits the behavior of AdvanceTicket while adjusting the pricing for students. If purchased 10 or more days before the event, student advance tickets cost $15, otherwise, $20, which is half the price of regular advance tickets. Additionally, the toString() method in this class adds an indication that a student ID is required for entry, providing complete ticket details. Overall, each class in the lab description serves a specific role in modeling different types of event tickets with their respective pricing rules and functionalities.
In the main class, we show the implementation of each class. The user is asked to input the number of tickets they are buying for each type of ticket. Then, the ticket number and price of the ticket is shown.