Create ArrayList.java
The class is named PhoneBookEntry. This class has two private variables ; name and phoneNumber. There are two accessor methods ; getName() and getPhoneNumber(); which returns the name and phoneNumber respectively. There are two mutator methods; setName() and setPhoneNumber() which sets the name and phoneNumber respectively. Lastly, there is a method ;displayentry(); which prints out the name and phoneNumber. 
I also created a class PhoneBookExample which implements the PhoneBookEntry class. In this class, I created five array list objects. The array list is named “phoneBook” After that, I declared a variable named “entry” of type PhoneBookEntry. This variable is used to reference each PhoneBookEntry object in the phoneBook. I used a for loop to display the name and phoneNumber of each PhoneBookEntry object. 
In order to make the program better, I changed the PhoneBookExample class so that the user inputs the name and phoneNumber of each new array list object. I also made sure that the phoneNumber that the user inputs consists of only 10 numbers. If the user inputs any less or any more than 10 numbers or any non-number input, the program asks the user to input the number again. A do-while loop was used for this purpose.