Create TowerOfHanoi.java
In the program there is a displayTowers() method that displays the current state of all three towers (A,B,C). It is called after each move to show the progression of the puzzle. The moveDisks() method uses a recursive method to solve the puzzle.If there is only one disk to move, it moves it from the “fromTower” to the “toTower”. If there are more than one disks, it ecursively moves n-1 disks from the fromTower to the auxTower using the toTower as an auxiliary.Then, it moves the nth disk from the fromTower to the toTower.After that, it recursively moves the n-1 disks from the auxTower to the toTower using the fromTower as an auxiliary. moveDisks() method focuses on the logic of solving the puzzle, while moveDisk() method handles the actual mechanics of moving a disk.In the moveDisk() method, the top disk from one twoer is moved to another. It uses stacks to push or pop the disks to the towers. 

Lastly, in the main() method, the number of disks are initialized and creates three stacks representing the towers. Then, it puts the disks in descending order, with the largest disk at the bottom in tower A. Then, the moveDisks() method is called.